Have tested enough of the display querying and initialization functionalities.
Have begun designing the new UI using wxFormBuilder. The tool will generate the class files for UI (The View) and I will create subclasses for each form that will contain the business logic. This will make it so that I can alter the GUI using wxFormBuilder and the business logic won't be erased.

More testing is necessary.

ADDITIONAL INFORMATION: New libraries are being introduced which should've been introduced when SDL and OpenGL were mentioned. GLEW and GLM are now part of the mix. See the VC++ include and lib directories to see where to extract your binaries / sources for the build to work.